refactor: share string literal unquoting helper
