+++ src/plugin/src/ast-transforms/consolidate-struct-assignments.js
@@ -9,6 +9,7 @@ import {
     getSingleMemberIndexPropertyEntry
 } from "../../../shared/ast-node-helpers.js";
 import { getCommentArray, isLineComment } from "../../../shared/comments.js";
+import { stripStringQuotes } from "../../../shared/string-utils.js";
 
 const FALLBACK_COMMENT_TOOLS = Object.freeze({
     addTrailingComment() {}
@@ -586,20 +587,6 @@ function isAttachableTrailingComment(comment, statement) {
     }
 
     return true;
-}
-
-function stripStringQuotes(value) {
-    if (typeof value !== "string" || value.length < 2) {
-        return null;
-    }
-
-    const firstChar = value[0];
-    const lastChar = value.at(-1);
-    if ((firstChar === '"' || firstChar === "'") && firstChar === lastChar) {
-        return value.slice(1, -1);
-    }
-
-    return null;
 }
 
 const IDENTIFIER_SAFE_PATTERN = /^[A-Za-z_][A-Za-z0-9_]*$/;
